Manejando errors, exceptions, headers y cookies

+++ db/db_user.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm.session import Session
 from db.models import DbUser
 from db.hash import Hash
@@ -23,13 +24,25 @@ def get_all_users(db: Session):
 
 
 def get_user_by_id(db: Session, id: int):
+    user = db.query(DbUser).filter(DbUser.id == id).first()
     # Handle any exceptions
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"User with id {id} not found",
+        )
+
     return db.query(DbUser).filter(DbUser.id == id).first()
 
 
 def update_user_by_id(db: Session, id: int, request: UserBase):
-    user = db.query(DbUser).filter(DbUser.id == id)
+    user = db.query(DbUser).filter(DbUser.id == id).first()
     # Handle any exceptions
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"User with id {id} not found",
+        )
     user.update(
         {
             DbUser.username: request.username,
@@ -44,6 +57,11 @@ def update_user_by_id(db: Session, id: int, request: UserBase):
 def delete_user_by_id(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
     # Handle any exceptions
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"User with id {id} not found",
+        )
     db.delete(user)
     db.commit()
     return "User deleted"
